requirements.txt
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
duckdb==1.2.2
fonttools==4.58.0
joblib==1.5.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyarrow==20.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
typer==0.15.3
typing_extensions==4.13.2
tzdata==2025.2
